[#2138] Refactor `D20Roll` for new API, convert `d20Roll` method
Rebuilds `D20Roll` to inherit from `BasicRoll`. This involved a
change to `BasicRoll` to give subclasses a way to modify message
data before the message is created.

As part of this process some of the logic that originally existed
in `D20Roll` has been moved to a new `D20Die` class. This includes
the getters for determining whether a roll is a critical success
or failure and applying advantage, minimums & maximums, and
halfling lucky. This change isolates the behavior that is specific
to rolling a d20 so that modules can potentially replace that die
with their own (e.g. 2d10 or 3d6).

This PR does not cover converting any of the existing rolling
methods over to the new API, they simply use the standalone
`d20Roll` method and let it handle calling the new API. Future PRs
will handle converting those methods over.