[#4061] Allow Cast Activity to trigger linked spells
Overrides the `use` method on `CastActivity` to trigger the cached
spell, rather than the activity itself. If the spell isn't cached,
it will first create a new cached copy. Triggers the `use` method
on the spell item so that the activity selection dialog can be
shown if necessary.

The `CastActivity` introduces two new hooks:
`dnd5e.preUseLinkedSpell` and `dnd5e.poseUseLinkedSpell`.

Adds a new option to `ActivtyUseConfiguration` called `triggering`
which contains the relative UUID of the activity triggering this
one as well as resource consumption options. This value is filled
automatically if the item has the `cachedFor` flag unless it is
explicitly set to `false`.

If `triggering` is set in the usage configuration, then during the
consumption process the system will also trigger the consumption
process on the triggering activity and merge its changes with that
of the main activity.

Some changes have been made to `ActivityUsageDialog` to show
consumption for both the triggering and triggered activities, and
to modify the spell slots list to reflect the scaling restriction
on the triggering activity.

Todo
- [ ] Properly handle consumption warnings on triggering activity
      once #4261 is resolved