[#2138, #3140] Re-implement concentration, death saves, & initiative
Re-implements `rollConcentration` and `rollDeathSave` using the new
rolling APIs, and modified `rollDeathSave` so it calls
`rollAbilitySave` rather than handling the roll itself.

Also reworked how initiative rolls are prepared, but due to the way
in which ability rolls interact with Foundry core these don't
fully support the new APIs. To better support this the `build`
method on `BasicRoll` can now take `evaluate: false` in the process
configuration, skipping the roll evaluation and message creation.
This allows the `rollInitiativeDialog` method to use `build` for
configuring its roll and calling the configuration hooks.

Added a `dnd5e.preConfigureInitiative` hook that is called for all
initiative rolls including those from the combat tracker. This
should allow some configuration, but doesn't allow canceling the
roll because core expects a roll to always be returned from
`getInitiativeRoll`.

Closes #3140