[#1530, #2138] Re-implement `rollSkill` and `rollToolCheck`
Merges the functionality ot skill & tool checks into a single
method that takes advantage of the new rolling APIs. These new
methods make use of a `buildConfig` callback that is called by
`RollConfigurationDialog` during its `buildConfig` method. This
allows for more extensive modifications of the configuration when
the ability is changed.

Closes #1530